+++ app/src/main/res/layout/activity_info.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimary"
-    tools:context=".InfoActivity">
+    tools:context=".ui.InfoActivity">
 
     <Button
         android:id="@+id/info_btn_back"
